requirements.txt
streamlit
langchain
langgraph
python-dotenv
pinecone-client  # Cliente para a base de dados vetorial Pinecone
